+++ test/abs/e2e/matrix_todo_test.go
@@ -1,9 +1,0 @@
-//go:build abs_e2e
-
-package e2e
-
-import "testing"
-
-func TestABSMatrixTODO_ABSMetadataModeLifecycle(t *testing.T) {
-	t.Fatal("TODO: build ABS metadata mode lifecycle tests")
-}